+++ CHANGELOG.md
@@ -16,6 +16,7 @@ TBD
 - Customise mask draw from interface builder [#50](https://github.com/JakeLin/IBAnimatable/issues/50) [#8](https://github.com/JakeLin/IBAnimatable/issues/8) [#76](https://github.com/JakeLin/IBAnimatable/issues/76):
   - Star(6) -> draw star mask with 6 sides.
   - Wave(up,50,100) -> draw a wave up mask with 50 as width, and 100 as offset.
+- `BorderDesignable` can be used directly in playground
 
 #### Bugfixes